+++ examples/s3-import/s3.tf
@@ -3,6 +3,7 @@ resource "aws_s3_bucket" "mysql" {
   #checkov:skip=CKV_AWS_19: Ensure all data stored in the S3 bucket is securely encrypted at rest
   #checkov:skip=CKV_AWS_144: Ensure that S3 bucket has cross-region replication enabled
   #checkov:skip=CKV_AWS_18: Ensure the S3 bucket has access logging enabled
+  #checkov:skip=CKV_AWS_145: "Ensure that S3 buckets are encrypted with KMS by default"
   bucket        = local.name
   force_destroy = true